abstract_model_inheritence_in_action.py
Here we created an abstract model called Planets and gave it fields like "name", "diameter", "moons". In Meta class we specified it as being abstract.
Just below this Planets model we created its child  model called Project and gave it few extra fields specific to this models itself, fields like "project_budget", "project_name".

In this way we created a robust model that is there to handle large amount of data.