Create migration_file.py
This code is about the structure of migration file, this holds a class name "Migration" a subclass of "migrations.Migration", This class defines two important attributes: 1) "dependencies" which is a list of other migrations that this one depends on. 2) "operations": A list of declarative instructions that specify the schema changes to be made.